Eliminate software double arithmetic from APU per-sample hot path
Both APU engines computed DC tracking using double precision:
  apu->dc_level_tracker += ((double)mixed - apu->dc_level_tracker) * 0.0001

On Xtensa LX7 (ESP32-S3), float is hardware-accelerated but double
requires software emulation (~250 cycles per operation vs ~5 for float).
With ~800 samples/frame, this cost ~0.9 ms per frame.

float precision is more than sufficient for this slow-moving (0.76 Hz)
DC offset tracker — the difference shows up only beyond 7 decimal places,
which is completely inaudible.

Expected improvement: ~0.9 ms/frame → gain ~2-3 fps toward 60 fps target.

+++ src/common/apu_infones.h
@@ -256,7 +256,7 @@ typedef struct {
 
     /* ---- Debug / metrics ---- */
     ApuDebugSampleStats channel_stats[APU_DEBUG_CHANNEL_COUNT];
-    double   dc_level_tracker;
+    float    dc_level_tracker;  // float is sufficient; double costs ~250 cycles/sample in SW on LX7
     uint32_t test_tone_phase[2];
 } Apu;
